# Fixes parsing http errors in connect client