--- a/.changeset/mighty-dragons-shake.md
+++ b/.changeset/mighty-dragons-shake.md
@@ -1,0 +1,5 @@
+---
+'@e2b/python-sdk': patch
+---
+
+fixes http error handling in e2b connect package